✨ Feat: AI 분ì석ã 로직 통합
