--- src/main/java/com/umc9th/bizscan/domain/aiAnalysis/dto/request/SwotAnalyzeRequest.java
@@ -1,15 +1,0 @@
-package com.umc9th.bizscan.domain.aiAnalysis.dto.request;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.AllArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class SwotAnalyzeRequest {
-
-    private Long storeId;
-}
