--- src/main/java/com/umc9th/bizscan/domain/aiAnalysis/dto/response/FastApiSwotResponse.java
@@ -1,21 +1,0 @@
-package com.umc9th.bizscan.domain.aiAnalysis.dto.response;
-
-import lombok.Getter;
-
-@Getter
-public class FastApiSwotResponse {
-
-    private String badge;
-
-    private String sTitle;
-    private String sDetail;
-
-    private String wTitle;
-    private String wDetail;
-
-    private String oTitle;
-    private String oDetail;
-
-    private String tTitle;
-    private String tDetail;
-}
